make fuzzy search an iterator
